# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Artel Orchestration Server API",
    "description": "API contract for SDK registration and agent action delivery through the Artel orchestration server.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the request"
    }
  ],
...
  "paths": {
...
    "/api/projects/{projectId}/documents": {
      "get": {
        "tags": [
          "Project Document"
        ],
        "summary": "기획서 이력",
        "description": "최신 버전이 앞에 온다.",
        "operationId": "list_11",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectDocumentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project Document"
        ],
        "summary": "기획서 등록",
        "description": "올라온 객체를 검증하고 새 버전으로 기록한다. 이전 버전은 남는다.",
        "operationId": "register_2",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/documents/upload-url": {
      "post": {
        "tags": [
          "Project Document"
        ],
        "summary": "업로드 URL 발급",
        "description": "PDF만 허용한다. 발급된 URL로 PUT할 때 세션 쿠키를 붙이면 서명이 깨진다.",
        "operationId": "createUploadTicket",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicketResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "프로젝트 상세",
        "description": "참여자가 아니면 404.",
        "operationId": "get_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "프로젝트 삭제",
        "description": "소유자만 가능하다. 실제로 지우지 않고 삭제 표시만 한다.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProjectResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Project"
        ],
        "summary": "프로젝트 수정",
        "description": "보낸 필드만 반영한다. 설명을 지우려면 빈 문자열을 보낸다.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/test-case-spec/download": {
      "get": {
        "tags": [
          "test-case-spec-controller"
        ],
        "operationId": "download",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestCaseSpecDownloadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/documents/{documentId}/download-url": {
      "get": {
        "tags": [
          "Project Document"
        ],
        "summary": "다운로드 URL 발급",
        "description": "요청할 때마다 새로 만드는 단기 URL이다.",
        "operationId": "createDownloadTicket",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "description": "기획서 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTicketResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/documents/events": {
      "get": {
        "tags": [
          "Project Document"
        ],
        "summary": "문서 추출 상태 스트림",
        "description": "프로젝트 문서 전부의 parse_status 변화를 SSE로 흘린다. 구독 직후 snapshot 프레임을 한 번 보내고, 그 뒤로는 parse_status가 바뀔 때마다 document 프레임을 보낸다. 접근할 수 없는 프로젝트는 stream이 열리지 않고 404로 끝난다.",
        "operationId": "events_2",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerSentEventDocumentStreamEvent"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/documents/{documentId}": {
      "delete": {
        "tags": [
          "Project Document"
        ],
        "summary": "기획서 삭제",
        "description": "이 버전을 지우고, 여기서 나온 knowledge도 함께 소프트삭제한다. 되돌릴 수 없다.",
        "operationId": "delete_5",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "description": "기획서 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DocumentUploaderResponse": {
        "required": [
          "displayName",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "ProjectDetailResponse": {
        "required": [
          "createdAt",
          "genre",
          "id",
          "myRole",
          "name",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "genre": {
            "type": "string",
            "enum": [
              "ACTION",
              "RPG",
              "PUZZLE",
              "SIMULATION",
              "STRATEGY",
              "SPORTS",
              "SHOOTER",
              "CASUAL",
              "OTHER"
            ]
          },
          "document": {
            "$ref": "#/components/schemas/ProjectDocumentResponse"
          },
          "myRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectDocumentResponse": {
        "required": [
          "contentType",
          "fileName",
          "id",
          "parseStatus",
          "sizeBytes",
          "uploadedAt",
          "uploadedBy",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "uploadedBy": {
            "$ref": "#/components/schemas/DocumentUploaderResponse"
          },
          "parseStatus": {
            "type": "string"
          }
        }
      },
...
      "RegisterDocumentRequest": {
        "required": [
          "objectKey"
        ],
        "type": "object",
        "properties": {
          "objectKey": {
            "type": "string"
          }
        }
      },
      "UploadTicketRequest": {
        "required": [
          "contentType",
          "fileName",
          "sizeBytes"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UploadTicketResponse": {
        "required": [
          "expiresAt",
          "objectKey",
          "requiredHeaders",
          "uploadUrl"
        ],
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "requiredHeaders": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "genre": {
            "type": "string",
            "enum": [
              "ACTION",
              "RPG",
              "PUZZLE",
              "SIMULATION",
              "STRATEGY",
              "SPORTS",
              "SHOOTER",
              "CASUAL",
              "OTHER"
            ]
          }
        }
      },
...
      "TestCaseSpecDownloadResponse": {
        "required": [
          "downloadUrl",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DownloadTicketResponse": {
        "required": [
          "downloadUrl",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ServerSentEventDocumentStreamEvent": {
        "type": "object"
      },
...
      "DeleteProjectResponse": {
        "required": [
          "deleted",
          "projectId"
        ],
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "projectId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}